climate-energy-analytics/docker/kafka_consumer/requirements.txt
kafka-python
boto3
pandas